steins-feed-tasks/pyproject.toml
[tool.poetry]
name = "steins-feed-tasks"
version = "0.1.0"
description = "Stein's feed -- tasks"
authors = ["Hans Yu <hans.yu@outlook.de>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
celery = {extras = ["redis"], version = "^5.4.0"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
